Ensure `.touch-border` works on small screens
The rules to float the element and collapse the gutter was previously only kicking in at the medium breakpoint. I've now moved the float out of this media query so it applies by default.

I also refactored the rule to move the media queries inside the selectors to remove the need to duplicate the rule selectors.