Add heading classes 1-6
I have purposefully added these classes in addition to the raw element selectors. While this restricted the ability to apply a `heading-1` class to a `h2` element, it will ensure backwards compatibility and that this change is not a large breaking change.

The larger sweep of these styles, decoupling them from raw element selectors will be included in time for v1.0

Fixes: #352