Convert !important overrides to sass-lint
The utilities need to use !important, so we need to override the standard rule.

Unfortunately, overrides in comments won't work in sass-lint until [this PR is merged](https://github.com/sasstools/sass-lint/pull/677), but these comments have been added anyway for when the feature is available.

Whenever this feature does become available, we should make the "no-important" rule in .sass-lint.yml strict.