Make default code blocks responsive
Unfortunately, scroll:auto on <pre> tags does not work as expected when the <pre> tag has a width of 100%, it needs an explicit value.

This means that by default we have to use white-space: pre-wrap to ensure we don’t break the layout on small screens where a long piece of code is displayed within a <pre> block.

Fixes #627