Move markdown docs back home
This is part of an experiment of putting Jekyll files at the top level, to make Docs live in a more sensible place. This will make docs more discoverable.

It should also signify to anyone familiar wit Jekyll that they can run Jekyll from this folder, which may also help with discoverability.