Reformat typography table markdown
So that the table will be treated predictably by different markdown
parsers, we should add preceding and following pipes (|) to each line.

In my opinion this also makes it look more like a table in text as well.

I've also normalised spacing, so it looks even more like a table.